+++ identity/src/test/java/com/stripe/android/identity/navigation/PassportUploadFragmentTest.kt
@@ -165,7 +165,7 @@ class PassportUploadFragmentTest {
     fun `upload front - kontinue enabled`() {
         launchFragment { binding, _, fragment ->
             runBlocking {
-                whenever(mockIdentityViewModel.postVerificationPageData(any(), any())).thenReturn(
+                whenever(mockIdentityViewModel.postVerificationPageData(any())).thenReturn(
                     VERIFICATION_PAGE_DATA_NOT_MISSING_BACK
                 )