+++ paymentsheet-example/src/androidTestDebug/java/com/stripe/android/screenshot/TestPaymentSheetScreenshots.kt
@@ -252,10 +252,11 @@ class TestPaymentSheetScreenshots {
         )
     }
 
-    companion object {
-        private val lpmRepository = LpmRepository(
-            LpmRepository.LpmRepositoryArguments(
-                InstrumentationRegistry.getInstrumentation().targetContext.resources
+    @Test
+    fun testPaymentSheetCustomPrimaryButtonLabel() {
+        testDriver.screenshotRegression(
+            testParams.copy(
+                customPrimaryButtonLabel = "Buy this now!",
             )
         )
     }